Update placements data type to be int

+++ src/Sfa.Tl.Matching.Domain/Models/Opportunity.cs
@@ -8,7 +8,7 @@ namespace Sfa.Tl.Matching.Domain.Models
         public short Distance { get; set; }
         public string JobTitle { get; set; }
         public bool? PlacementsKnown { get; set; }
-        public short? Placements { get; set; }
+        public int? Placements { get; set; }
         public string EmployerName { get; set; }
         public string EmployerContact { get; set; }
         public string EmployerContactEmail { get; set; }
